[Unified Observability] Add feature flag for the new overview page (#119193)
* Add feature flag to display a blank overview page when enabled

* Add tests for overview page feature flag

* Fix types

* Fix more types

* Remove duplicated BucketSize type

* fix linter
