+++ x-pack/plugins/observability/public/components/app/section/ux/index.test.tsx
@@ -42,7 +42,13 @@ describe('UXSection', () => {
         http: { basePath: { prepend: jest.fn() } },
       } as unknown as CoreStart,
       appMountParameters: {} as AppMountParameters,
-      config: { unsafe: { alertingExperience: { enabled: true }, cases: { enabled: true } } },
+      config: {
+        unsafe: {
+          alertingExperience: { enabled: true },
+          cases: { enabled: true },
+          overviewNext: { enabled: false },
+        },
+      },
       plugins: {
         data: {
           query: {
