+++ x-pack/plugins/observability/public/hooks/use_time_range.test.ts
@@ -24,7 +24,13 @@ describe('useTimeRange', () => {
     jest.spyOn(pluginContext, 'usePluginContext').mockImplementation(() => ({
       core: {} as CoreStart,
       appMountParameters: {} as AppMountParameters,
-      config: { unsafe: { alertingExperience: { enabled: true }, cases: { enabled: true } } },
+      config: {
+        unsafe: {
+          alertingExperience: { enabled: true },
+          cases: { enabled: true },
+          overviewNext: { enabled: false },
+        },
+      },
       plugins: {
         data: {
           query: {
@@ -67,7 +73,13 @@ describe('useTimeRange', () => {
         jest.spyOn(pluginContext, 'usePluginContext').mockImplementation(() => ({
           core: {} as CoreStart,
           appMountParameters: {} as AppMountParameters,
-          config: { unsafe: { alertingExperience: { enabled: true }, cases: { enabled: true } } },
+          config: {
+            unsafe: {
+              alertingExperience: { enabled: true },
+              cases: { enabled: true },
+              overviewNext: { enabled: false },
+            },
+          },
           plugins: {
             data: {
               query: {
