+++ x-pack/plugins/observability/public/index.ts
@@ -26,7 +26,11 @@ export type {
 export { enableInspectEsQueries } from '../common/ui_settings_keys';
 
 export interface ConfigSchema {
-  unsafe: { alertingExperience: { enabled: boolean }; cases: { enabled: boolean } };
+  unsafe: {
+    alertingExperience: { enabled: boolean };
+    cases: { enabled: boolean };
+    overviewNext: { enabled: boolean };
+  };
 }
 
 export const plugin: PluginInitializer<
