+++ x-pack/plugins/observability/public/pages/overview/overview.stories.tsx
@@ -60,7 +60,11 @@ const withCore = makeDecorator({
                 setHeaderActionMenu: () => {},
               } as unknown as AppMountParameters,
               config: {
-                unsafe: { alertingExperience: { enabled: true }, cases: { enabled: true } },
+                unsafe: {
+                  alertingExperience: { enabled: true },
+                  cases: { enabled: true },
+                  overviewNext: { enabled: false },
+                },
               },
               core: options as CoreStart,
               plugins: {
